refactor(rdb): group function by resource
