--- internal/namespaces/rdb/v1/custom_instance_create_test.go
@@ -1,17 +1,0 @@
-package rdb
-
-import (
-	"fmt"
-	"testing"
-
-	"github.com/scaleway/scaleway-cli/internal/core"
-)
-
-func Test_CreateInstance(t *testing.T) {
-	t.Run("Simple", core.Test(&core.TestConfig{
-		Commands:  GetCommands(),
-		Cmd:       fmt.Sprintf("scw rdb instance create node-type=DB-DEV-S is-ha-cluster=false name=%s engine=%s user-name=%s password=%s --wait", name, engine, user, password),
-		Check:     core.TestCheckGolden(),
-		AfterFunc: core.ExecAfterCmd("scw rdb instance delete {{ .CmdResult.ID }}"),
-	}))
-}
