+++ internal/namespaces/rdb/v1/custom_instance_test.go
@@ -1,0 +1,47 @@
+package rdb
+
+import (
+	"fmt"
+	"testing"
+
+	"github.com/scaleway/scaleway-cli/internal/core"
+)
+
+func Test_CloneInstance(t *testing.T) {
+	t.Run("Simple", core.Test(&core.TestConfig{
+		Commands:   GetCommands(),
+		BeforeFunc: createInstance(),
+		Cmd:        "scw rdb instance clone {{ .Instance.ID }} node-type=DB-DEV-M name=foobar --wait",
+		Check:      core.TestCheckGolden(),
+		AfterFunc:  deleteInstance(),
+	}))
+}
+
+func Test_CreateInstance(t *testing.T) {
+	t.Run("Simple", core.Test(&core.TestConfig{
+		Commands:  GetCommands(),
+		Cmd:       fmt.Sprintf("scw rdb instance create node-type=DB-DEV-S is-ha-cluster=false name=%s engine=%s user-name=%s password=%s --wait", name, engine, user, password),
+		Check:     core.TestCheckGolden(),
+		AfterFunc: core.ExecAfterCmd("scw rdb instance delete {{ .CmdResult.ID }}"),
+	}))
+}
+
+func Test_GetInstance(t *testing.T) {
+	t.Run("Simple", core.Test(&core.TestConfig{
+		Commands:   GetCommands(),
+		BeforeFunc: createInstance(),
+		Cmd:        "scw rdb instance get {{ .StartServer.ID }}",
+		Check:      core.TestCheckGolden(),
+		AfterFunc:  deleteInstance(),
+	}))
+}
+
+func Test_UpgradeInstance(t *testing.T) {
+	t.Run("Simple", core.Test(&core.TestConfig{
+		Commands:   GetCommands(),
+		BeforeFunc: createInstance(),
+		Cmd:        "scw rdb instance upgrade {{ .StartServer.ID }} node-type=DB-DEV-M --wait",
+		Check:      core.TestCheckGolden(),
+		AfterFunc:  deleteInstance(),
+	}))
+}
